Kill existing go-eldoc buffer to prevent duplicates
Previously when using any of the gocode variants and rapidly
triggering go-eldoc to show eldoc info in the modebar, if the context
was killed by gocode this would result in multiple go-eldoc
buffers. This commit now ensures that if an initial go-eldoc buffer
exists, it is killed before the new temp-buffer is created.